requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.1.0
    # via streamlit
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   streamlit
    #   uvicorn
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
dnspython==2.7.0
    # via pymongo
executing==2.2.0
    # via stack-data
fastapi==0.115.14
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.58.4
    # via matplotlib
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.16.0
    # via uvicorn
hf-xet==1.1.5
    # via huggingface-hub
huggingface-hub==0.33.1
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.1.0
    # via pytest
ipython==9.3.0
    # via -r requirements.in
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.1
    # via scikit-learn
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via
    #   -r requirements.in
    #   wordcloud
matplotlib-inline==0.1.7
    # via ipython
mpmath==1.3.0
    # via sympy
narwhals==1.44.0
    # via
    #   altair
    #   plotly
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   transformers
    #   wordcloud
oauthlib==3.3.1
    # via
    #   requests-oauthlib
    #   tweepy
packaging==25.0
    # via
    #   altair
    #   huggingface-hub
    #   matplotlib
    #   plotly
    #   pytest
    #   streamlit
    #   transformers
pandas==2.3.0
    # via
    #   -r requirements.in
    #   streamlit
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via
    #   matplotlib
    #   streamlit
    #   wordcloud
plotly==6.2.0
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.51
    # via ipython
protobuf==6.31.1
    # via streamlit
psycopg2-binary==2.9.10
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==20.0.0
    # via streamlit
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
pymongo==4.13.2
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   requests-oauthlib
    #   streamlit
    #   transformers
    #   tweepy
requests-oauthlib==2.0.0
    # via tweepy
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via -r requirements.in
scipy==1.16.0
    # via scikit-learn
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via -r requirements.in
stack-data==0.6.3
    # via ipython
starlette==0.46.2
    # via fastapi
streamlit==1.46.1
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.7.1
    # via -r requirements.in
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
transformers==4.53.0
    # via -r requirements.in
tweepy==4.16.0
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   altair
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   ipython
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
wordcloud==1.9.4
    # via -r requirements.in
